Add files
This a simplified version of the code I've been using to visualise text data.  It should be possible to visualise a field of study with this code.  Currently the colouring of the output varies with citation counts to help you find the best cited topics in the field.  (Note that it is possible to change the response variable to any of Dimensions' new citation metrics.  E.g. RCR or FCR.)